Fixed api

+++ museumback/controllers/categoryController.ts
@@ -3,28 +3,85 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 import { validateCategoryId, validateImage, validateMuseumId, validateMuseumIdCategoryId, validatePagination } from "../utils/validations";
 import { uploadFile } from "../lib/storage";
+import { getCurrentUser } from "../middleware/authenticated";
+import { TokenPayload } from "../utils/tokenator";
+import { validatedMuseumAndCategory } from "./itemController";
+
+export async function validatedMuseum(museumId: number, res: Response): Promise<boolean> {
+    const mus = await prisma.museum.findFirst({
+        where: {
+            id: museumId
+        }
+    });
+    if (!mus) {
+        res.status(404).json({ errors: [`Museum not found!`] });
+        return false;
+    }
+    return true;
+}
 
 export async function searchCategories(req: Request, res: Response, next: NextFunction) {
     try {
-        const query = z.object({
+        const query = validatePagination.merge(z.object({
             name: z.string().default(''),
-        }).parse(req.query);
+        })).parse(req.query);
         const params = validateMuseumId.parse(req.params);
-        const categories = await prisma.category.findMany({
-            take: 5,
-            where: {
-                name: {
-                    contains: query.name,
-                    mode: 'insensitive'
-                },
-                museumId: params.museumId
-            },
-            select: {
-                name: true,
-                id: true
-            }
-        })
-        return res.status(200).json({ categories });
+        // validating if museum exists
+        if (!await validatedMuseum(params.museumId, res))
+            return;
+        const user = res.locals.user as TokenPayload;
+        let categories = []
+        switch (user.role) {
+            case 'CURATOR':
+                categories = await prisma.category.findMany({
+                    take: query.pageSize + 1,
+                    skip: (query.page - 1) * query.pageSize,
+                    where: {
+                        UserCategory: {
+                            some: {
+                                userId: user.id,
+                                category: {
+                                    museumId: params.museumId
+                                }
+                            }
+                        }
+                    },
+                    select: {
+                        name: true,
+                        id: true,
+                        description: true,
+                        imageUrl: true
+                    }
+                });
+                break;
+            case 'ADMIN':
+                categories = await prisma.category.findMany({
+                    take: query.pageSize + 1,
+                    skip: (query.page - 1) * query.pageSize,
+                    where: {
+                        name: {
+                            contains: query.name,
+                            mode: 'insensitive'
+                        },
+                        museumId: params.museumId
+                    },
+                    select: {
+                        name: true,
+                        id: true,
+                        description: true,
+                        imageUrl: true
+                    }
+                });
+                break;
+            default:
+                return res.status(501).json({ errors: [`Category search for user role '${user.role}' is not implemented!`] });
+        }
+        let hasNext = false;
+        if (categories.length - 1 === query.pageSize) {
+            hasNext = true;
+            categories.pop();
+        }
+        return res.status(200).json({ categories, hasNext });
     } catch (err) {
         next(err);
     }
@@ -34,7 +91,9 @@ export async function getCategories(req: Request, res: Response, next: NextFunct
     try {
         const params = validateMuseumId.parse(req.params);
         const query = validatePagination.parse(req.query);
-
+        // validating if museum exists
+        if (!await validatedMuseum(params.museumId, res))
+            return;
         const categories = await prisma.category.findMany({
             take: query.pageSize + 1,
             skip: (query.page - 1) * query.pageSize,
@@ -56,6 +115,9 @@ export async function getCategories(req: Request, res: Response, next: NextFunct
 export async function getCategory(req: Request, res: Response, next: NextFunction) {
     try {
         const params = validateMuseumIdCategoryId.parse(req.params);
+        // validating if museum exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
         const category = await prisma.category.findFirst({
             where: {
                 museumId: params.museumId,
@@ -66,11 +128,29 @@ export async function getCategory(req: Request, res: Response, next: NextFunctio
                     select: {
                         name: true
                     }
-                }
+                },
             }
         });
         if (!category) return res.status(404).json({ errors: ['Category is not found'] });
-        return res.status(200).json({ category });
+        const user = await getCurrentUser(req);
+        let canEdit = false;
+        if (user) {
+            switch (user.role) {
+                case 'ADMIN':
+                    canEdit = true;
+                    break;
+                case 'CURATOR':
+                    const userCat = await prisma.userCategory.findFirst({
+                        where: {
+                            userId: user.id,
+                            categoryId: params.categoryId
+                        }
+                    })
+                    if (userCat) canEdit = true;
+                    break;
+            }
+        }
+        return res.status(200).json({ category, canEdit });
     } catch (rr) {
         next(rr);
     }
@@ -84,6 +164,9 @@ export async function createCategory(req: Request, res: Response, next: NextFunc
     try {
         const params = validateMuseumId.parse(req.params);
         const body = validateCreateParams.parse(req.fields);
+        // validating if museum exists
+        if (!await validatedMuseum(params.museumId, res))
+            return;
         const mus = await prisma.museum.findFirst({
             where: {
                 id: params.museumId
@@ -110,16 +193,9 @@ export async function updateCategory(req: Request, res: Response, next: NextFunc
     try {
         const params = validateMuseumIdCategoryId.parse(req.params);
         const body = validateCreateParams.merge(validateMuseumId).parse(req.fields);
-        // check if the current museum exists
-        const mus = await prisma.museum.findFirst({
-            where: {
-                id: params.museumId
-            },
-            select: {
-                id: true
-            }
-        });
-        if (!mus) return res.status(404).json({ errors: ['Museum not found!'] });
+        // validating if museum exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
         // check if new museum exists
         const newMus = await prisma.museum.findFirst({
             where: {
@@ -141,6 +217,7 @@ export async function updateCategory(req: Request, res: Response, next: NextFunc
                 museumId: body.museumId
             }
         });
+
         return res.status(200).json({ category });
     } catch (rr) {
         next(rr);
@@ -149,7 +226,10 @@ export async function updateCategory(req: Request, res: Response, next: NextFunc
 
 export async function deleteCategory(req: Request, res: Response, next: NextFunction) {
     try {
-        const params = validateCategoryId.parse(req.params);
+        const params = validateMuseumIdCategoryId.parse(req.params);
+        // validating if museum exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
         await prisma.category.delete({
             where: {
                 id: params.categoryId
@@ -165,6 +245,9 @@ export async function updateImage(req: Request, res: Response, next: NextFunctio
     try {
         const params = validateMuseumIdCategoryId.parse(req.params);
         const files = validateImage.parse(req.files);
+        // validating if museum and category exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
         if (!files.image) {
             return res.status(400).json({ errors: [`Missing image file`] });
         }
@@ -178,6 +261,103 @@ export async function updateImage(req: Request, res: Response, next: NextFunctio
                 imageUrl: url
             }
         })
+        return res.status(204).send();
+    } catch (err) {
+        next(err);
+    }
+}
+
+export async function getUsers(req: Request, res: Response, next: NextFunction) {
+    try {
+        const params = validateMuseumIdCategoryId.parse(req.params);
+        // validating if museum exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
+        const users = await prisma.userCategory.findMany({
+            where: {
+                categoryId: params.categoryId,
+            },
+            select: {
+                user: {
+                    select: {
+                        id: true,
+                        email: true
+                    }
+                }
+            }
+        });
+        const formattedUsers = users.map(w => ({
+            id: w.user.id,
+            email: w.user.email
+        }))
+        return res.status(200).json({ users: formattedUsers });
+    } catch (err) {
+        next(err);
+    }
+}
+
+const validateUserId = z.object({
+    id: z.number().min(0)
+})
+
+export async function addUser(req: Request, res: Response, next: NextFunction) {
+    try {
+        const params = validateMuseumIdCategoryId.parse(req.params);
+        const body = validateUserId.parse(req.fields);
+        // validating if museum and category exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
+        const user = await prisma.user.findFirst({
+            where: {
+                id: body.id
+            }
+        });
+        if (!user) return res.status(404).json({ errors: ['User not found'] });
+        if (user.role !== 'CURATOR') return res.status(403).json({ errors: [`Can't add non curator user to a category!`] });
+        const userCat = await prisma.userCategory.findFirst({
+            where: {
+                categoryId: params.categoryId,
+                userId: body.id
+            }
+        })
+        if (userCat) return res.status(409).json({ errors: ['User is already added to this category!'] });
+        await prisma.userCategory.create({
+            data: {
+                userId: body.id,
+                categoryId: params.categoryId,
+                assignorId: res.locals.user.id
+            }
+        });
+        return res.status(204).send();
+    } catch (err) {
+        next(err);
+    }
+}
+
+export async function removeUser(req: Request, res: Response, next: NextFunction) {
+    try {
+        const params = validateMuseumIdCategoryId.merge(z.object({
+            id: z.coerce.number().min(0)
+        })).parse(req.params);
+
+        // validating if museum and category exists
+        if (!await validatedMuseumAndCategory(params.museumId, params.categoryId, res))
+            return;
+        const userCat = await prisma.userCategory.findFirst({
+            where: {
+                categoryId: params.categoryId,
+                userId: params.id,
+            }
+        })
+        if (!userCat) return res.status(404).json({ errors: ['User is not found on that category!'] });
+        await prisma.userCategory.delete({
+            where: {
+                userId_categoryId: {
+                    userId: params.id,
+                    categoryId: params.categoryId
+                }
+            }
+        });
         return res.status(204).send();
     } catch (err) {
         next(err);
